SimpleOpenCLGenerator -> FloatOpenCLGenerator dodanie DoubleOpenCLGenerator przej?cie na double

+++ Mandelbrot/MultiThreadedGenerator.cs
@@ -8,11 +8,11 @@ namespace Baron.Mandelbrot
 {
     public class MultiThreadedGenerator : IGenerator
     {
-        public byte[] Generate(int aWidth, int aHeight, float aSetStartX, float aSetWidth, float aSetStartY, float aSetHeight, byte[] aPalette)
+        public byte[] Generate(int aWidth, int aHeight, double aSetStartX, double aSetWidth, double aSetStartY, double aSetHeight, byte[] aPalette)
         {
             byte[] rgbaValues = new byte[aWidth * aHeight * 4];
-            float x_step = aSetWidth / aWidth;
-            float y_step = aSetHeight / aHeight;
+            double x_step = aSetWidth / aWidth;
+            double y_step = aSetHeight / aHeight;
 
             Parallel.For(0, aHeight, screenY =>
             {
@@ -31,16 +31,16 @@ namespace Baron.Mandelbrot
             return rgbaValues;
         }
 
-        private byte GetIteration(float aSetStartX, float aSetStartY, float x_step, float y_step, int screenY, int screenX)
+        private byte GetIteration(double aSetStartX, double aSetStartY, double x_step, double y_step, int screenY, int screenX)
         {
-            float x0 = aSetStartX + screenX * x_step;
-            float y0 = aSetStartY + screenY * y_step;
+            double x0 = aSetStartX + screenX * x_step;
+            double y0 = aSetStartY + screenY * y_step;
             byte iteration = 0;
-            float x = 0;
-            float y = 0;
+            double x = 0;
+            double y = 0;
             while (x * x + y * y <= 4 && iteration < 255)
             {
-                float tempX = x * x - y * y + x0;
+                double tempX = x * x - y * y + x0;
                 y = 2 * x * y + y0;
                 x = tempX;
                 iteration++;
